Move Python backend implementation into main module
